sdk-macro: Move package_metadata macro to separate crate (#1713)
* sdk-macro: Move package_metadata macro to separate crate

* Update sbf lockfile

* Cargo fmt
