+++ Cargo.toml
@@ -97,6 +97,7 @@ members = [
     "sdk/cargo-test-sbf",
     "sdk/gen-headers",
     "sdk/macro",
+    "sdk/package-metadata-macro",
     "sdk/program",
     "send-transaction-service",
     "stake-accounts",
@@ -359,6 +360,7 @@ solana-metrics = { path = "metrics", version = "=2.0.0" }
 solana-net-utils = { path = "net-utils", version = "=2.0.0" }
 solana-nohash-hasher = "0.2.1"
 solana-notifier = { path = "notifier", version = "=2.0.0" }
+solana-package-metadata-macro = { path = "sdk/package-metadata-macro", version = "=2.0.0" }
 solana-perf = { path = "perf", version = "=2.0.0" }
 solana-poh = { path = "poh", version = "=2.0.0" }
 solana-poseidon = { path = "poseidon", version = "=2.0.0" }
