+++ sdk/package-metadata-macro/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "solana-sdk-macro"
-description = "Solana SDK Macro"
-documentation = "https://docs.rs/solana-sdk-macro"
+name = "solana-package-metadata-macro"
+description = "Solana Package Metadata Macro"
+documentation = "https://docs.rs/solana-package-metadata-macro"
 version = { workspace = true }
 authors = { workspace = true }
 repository = { workspace = true }
@@ -13,11 +13,10 @@ edition = { workspace = true }
 proc-macro = true
 
 [dependencies]
-bs58 = { workspace = true }
 proc-macro2 = { workspace = true }
 quote = { workspace = true }
-rustversion = { workspace = true }
 syn = { workspace = true, features = ["full"] }
+toml = { workspace = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
